# api/docs/openapi.yaml
openapi: 3.0.3
paths:
...
  /organizations/{organization_uuid}/ai-providers:
    get:
      operationId: AiProvidersController_findAll
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - AiProviders
      security:
        - bearer: []
    post:
      operationId: AiProvidersController_create
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAiProviderDto"
      responses:
        "201":
          description: ""
      tags:
        - AiProviders
      security:
        - bearer: []
  /organizations/{organization_uuid}/ai-providers/{provider_uuid}:
    patch:
      operationId: AiProvidersController_update
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: provider_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAiProviderDto"
      responses:
        "200":
          description: ""
      tags:
        - AiProviders
      security:
        - bearer: []
    delete:
      operationId: AiProvidersController_delete
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: provider_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - AiProviders
      security:
        - bearer: []
...
  /organizations/{organization_uuid}/integrations/apps/connect:
    post:
      operationId: ComposioConnectionsController_connect[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectComposioDto"
      responses:
        "201":
          description: ""
      tags:
        - ComposioConnections
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/apps/callback:
    post:
      operationId: ComposioConnectionsController_verifyCallback[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComposioCallbackDto"
      responses:
        "201":
          description: ""
      tags:
        - ComposioConnections
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/apps/accounts:
    get:
      operationId: ComposioConnectionsController_listAccounts[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: toolkit_slug
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - ComposioConnections
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/apps/accounts/{connected_account_id}:
    delete:
      operationId: ComposioConnectionsController_disconnect[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: connected_account_id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - ComposioConnections
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/apps/accounts/{connected_account_id}/reconnect:
    post:
      operationId: ComposioConnectionsController_reconnect[0]
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: connected_account_id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - ComposioConnections
      security:
        - bearer: []
...
  /organizations/{organization_uuid}/agents:
    get:
      operationId: AgentsController_findAll
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Agents
      security:
        - bearer: []
    post:
      operationId: AgentsController_create
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentDto"
      responses:
        "201":
          description: ""
      tags:
        - Agents
      security:
        - bearer: []
  /organizations/{organization_uuid}/agents/{agent_uuid}:
    get:
      operationId: AgentsController_findOne
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: agent_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Agents
      security:
        - bearer: []
    patch:
      operationId: AgentsController_update
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: agent_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentDto"
      responses:
        "200":
          description: ""
      tags:
        - Agents
      security:
        - bearer: []
    delete:
      operationId: AgentsController_remove
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: agent_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Agents
      security:
        - bearer: []
info:
  title: Cortex API
  description: >-
    REST API for Cortex — organizations, conversations, agents, integrations (Composio, MCP, OpenAPI, databases),
    documents, roles, and members. Authenticate via JWT Bearer token from POST /auth/email/login.
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3001
    description: Local
  - url: "{{baseUrl}}"
    description: Environment variable
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /auth/email/login
  schemas:
...
    CreateAiProviderDto:
      type: object
      properties:
        provider:
          type: string
          enum:
            - OPENAI
            - CLAUDE
            - GROK
        api_key:
          type: string
        default_model:
          type: string
        model_routing:
          type: object
        usage_limit_tokens:
          type: number
        usage_limit_cost_usd:
          type: number
        is_default:
          type: boolean
      required:
        - provider
        - api_key
        - default_model
    UpdateAiProviderDto:
      type: object
      properties:
        provider:
          type: string
          enum:
            - OPENAI
            - CLAUDE
            - GROK
        api_key:
          type: string
        default_model:
          type: string
        model_routing:
          type: object
        usage_limit_tokens:
          type: number
        usage_limit_cost_usd:
          type: number
        is_default:
          type: boolean
...
    ConnectComposioDto:
      type: object
      required:
        - toolkit_slug
      properties:
        toolkit_slug:
          type: string
        connected_account_id:
          type: string
        connection_tier:
          type: string
          enum:
            - ORG_SHARED
            - USER_PERSONAL
    ComposioCallbackDto:
      type: object
      required:
        - toolkit_slug
      properties:
        toolkit_slug:
          type: string
        connection_request_id:
          type: string
        connected_account_id:
          type: string
        connection_tier:
          type: string
          enum:
            - ORG_SHARED
            - USER_PERSONAL
...
    CreateAgentDto:
      type: object
      properties:
        title:
          type: string
          example: Daily sales summary
        prompt:
          type: string
          example: Summarize yesterday sales and highlight anomalies.
        cron_expression:
          type: string
          example: 0 9 * * *
          description: 5-field cron expression
        is_enabled:
          type: boolean
          default: true
      required:
        - title
        - prompt
        - cron_expression
    UpdateAgentDto:
      type: object
      properties:
        title:
          type: string
          example: Daily sales summary
        prompt:
          type: string
          example: Summarize yesterday sales and highlight anomalies.
        cron_expression:
          type: string
          example: 0 9 * * *
          description: 5-field cron expression
        is_enabled:
          type: boolean
          default: true
security:
  - bearer: []